Restructure + ALLINONE Main file. + GPS LED blink for Teensy
Restructured so you don't have to worry about headers being in different folders.

The tests folder will hold all of the separate ino files for each test we do on a separate breadboard.

In addition, I created the Main_ALLINONE.ino file which will probably be the go to so we don't have to deal with headers and including proper files

NOTE: I don't believe that the radio will output the right thing as well as possibly mess with the GUI.